Move test code out, write a strongly typed tutorial in App/. Need to fix checkbox values.
