+++ packages/formik/src/hooks/useTypedField.tsx
@@ -9,10 +9,20 @@ export type TypedField<FormValues> = <
 ) =>
   React.ReactElement | null;
 
+/**
+ * Create a typed field from anywhere.
+ */
+export const createTypedField = <Values,>(): TypedField<Values> => (
+  props
+) =>
+  <Field {...props} />;
+
+/**
+ * Create a TypedField from within Formik.
+ *
+ * @private
+ */
 export const useTypedField = <Values,>() => React.useMemo<TypedField<Values>>(
-    () => (
-      props
-    ) =>
-      <Field {...props} />,
+    () => createTypedField(),
     []
 );
