+++ packages/formik/src/hooks/useTypedFieldArray.tsx
@@ -6,11 +6,17 @@ export type TypedFieldArray<FormValues> = <Path extends string>(
 ) =>
   React.ReactElement | null;
 
+/**
+ * Get TypedFieldArray from anywhere.
+ */
+export const createTypedFieldArray = <Values,>(): TypedFieldArray<Values> =>
+  (props) => <FieldArray {...props} />;
+
+/**
+ * Use TypedFieldArray from within a component.
+ */
 export const useTypedFieldArray = <Values,>() =>
   React.useMemo<TypedFieldArray<Values>>(
-    () => (
-      props
-    ) =>
-      <FieldArray {...props} />,
+    () => createTypedFieldArray(),
     []
   );
